+++ packages/orq-rc/src/orq_ai_sdk/feedback.py
@@ -80,6 +80,7 @@ class Feedback(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateFeedback",
                 oauth2_scopes=[],
@@ -187,6 +188,7 @@ class Feedback(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateFeedback",
                 oauth2_scopes=[],