+++ packages/orq-rc/src/orq_ai_sdk/knowledge.py
@@ -77,6 +77,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListKnowledgeBases",
                 oauth2_scopes=[],
@@ -181,6 +182,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListKnowledgeBases",
                 oauth2_scopes=[],
@@ -296,6 +298,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateKnowledge",
                 oauth2_scopes=[],
@@ -411,6 +414,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateKnowledge",
                 oauth2_scopes=[],
@@ -509,6 +513,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOneKnowledge",
                 oauth2_scopes=[],
@@ -607,6 +612,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOneKnowledge",
                 oauth2_scopes=[],
@@ -732,6 +738,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateKnowledge",
                 oauth2_scopes=[],
@@ -857,6 +864,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateKnowledge",
                 oauth2_scopes=[],
@@ -955,6 +963,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteKnowledge",
                 oauth2_scopes=[],
@@ -1051,6 +1060,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteKnowledge",
                 oauth2_scopes=[],
@@ -1177,6 +1187,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="SearchKnowledge",
                 oauth2_scopes=[],
@@ -1305,6 +1316,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="SearchKnowledge",
                 oauth2_scopes=[],
@@ -1416,6 +1428,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListDatasources",
                 oauth2_scopes=[],
@@ -1527,6 +1540,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListDatasources",
                 oauth2_scopes=[],
@@ -1645,6 +1659,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateDatasource",
                 oauth2_scopes=[],
@@ -1763,6 +1778,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateDatasource",
                 oauth2_scopes=[],
@@ -1862,6 +1878,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="RetrieveDatasource",
                 oauth2_scopes=[],
@@ -1961,6 +1978,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="RetrieveDatasource",
                 oauth2_scopes=[],
@@ -2062,6 +2080,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteDatasource",
                 oauth2_scopes=[],
@@ -2161,6 +2180,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteDatasource",
                 oauth2_scopes=[],
@@ -2270,6 +2290,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateDatasource",
                 oauth2_scopes=[],
@@ -2381,6 +2402,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateDatasource",
                 oauth2_scopes=[],
@@ -2494,6 +2516,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateChunk",
                 oauth2_scopes=[],
@@ -2607,6 +2630,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreateChunk",
                 oauth2_scopes=[],
@@ -2723,6 +2747,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListChunks",
                 oauth2_scopes=[],
@@ -2839,6 +2864,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListChunks",
                 oauth2_scopes=[],
@@ -2964,6 +2990,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateChunk",
                 oauth2_scopes=[],
@@ -3089,6 +3116,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdateChunk",
                 oauth2_scopes=[],
@@ -3191,6 +3219,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteChunk",
                 oauth2_scopes=[],
@@ -3291,6 +3320,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeleteChunk",
                 oauth2_scopes=[],
@@ -3391,6 +3421,7 @@ class Knowledge(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOneChunk",
                 oauth2_scopes=[],
@@ -3493,6 +3524,7 @@ class Knowledge(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOneChunk",
                 oauth2_scopes=[],