+++ packages/orq-rc/src/orq_ai_sdk/metrics.py
@@ -120,6 +120,7 @@ class Metrics(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeploymentCreateMetric",
                 oauth2_scopes=[],
@@ -267,6 +268,7 @@ class Metrics(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeploymentCreateMetric",
                 oauth2_scopes=[],