+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-29T12:34:18.367Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-30T11:01:20.758Z")
     r"""The date and time the resource was last updated"""