+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-29T12:34:21.136Z")
+    created: Optional[datetime] = parse_datetime("2025-05-30T11:01:23.565Z")
     r"""The date and time the resource was created"""