+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWE0JQ8E7P71NHV2H1D5899R"
+        "01JWGDN7FP860EXPKS0634AF99"
     )
     r"""The id of the resource"""