+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWE0JQ8H1ZFGBWXQTHYW8Z0R"
+        "01JWGDN7FSZNZJWHH2HRE5V7XJ"
     )
     r"""The id of the resource"""