+++ packages/orq-rc/src/orq_ai_sdk/models_.py
@@ -64,6 +64,7 @@ class Models(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListModels",
                 oauth2_scopes=[],
@@ -155,6 +156,7 @@ class Models(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListModels",
                 oauth2_scopes=[],