+++ packages/orq-rc/src/orq_ai_sdk/prompts.py
@@ -77,6 +77,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetAllPrompts",
                 oauth2_scopes=[],
@@ -181,6 +182,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetAllPrompts",
                 oauth2_scopes=[],
@@ -284,6 +286,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreatePrompt",
                 oauth2_scopes=[],
@@ -387,6 +390,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="CreatePrompt",
                 oauth2_scopes=[],
@@ -485,6 +489,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOnePrompt",
                 oauth2_scopes=[],
@@ -583,6 +588,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetOnePrompt",
                 oauth2_scopes=[],
@@ -732,6 +738,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdatePrompt",
                 oauth2_scopes=[],
@@ -887,6 +894,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="UpdatePrompt",
                 oauth2_scopes=[],
@@ -989,6 +997,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeletePrompt",
                 oauth2_scopes=[],
@@ -1083,6 +1092,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="DeletePrompt",
                 oauth2_scopes=[],
@@ -1188,6 +1198,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListPromptVersions",
                 oauth2_scopes=[],
@@ -1295,6 +1306,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="ListPromptVersions",
                 oauth2_scopes=[],
@@ -1396,6 +1408,7 @@ class Prompts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetPromptVersion",
                 oauth2_scopes=[],
@@ -1503,6 +1516,7 @@ class Prompts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="GetPromptVersion",
                 oauth2_scopes=[],