+++ packages/orq-rc/src/orq_ai_sdk/remoteconfigs.py
@@ -83,6 +83,7 @@ class Remoteconfigs(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="RemoteConfigsGetConfig",
                 oauth2_scopes=[],
@@ -193,6 +194,7 @@ class Remoteconfigs(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="RemoteConfigsGetConfig",
                 oauth2_scopes=[],